==> 00_setup.ipynb <==
import os
import sys
import math
import warnings
import yaml
import pandas as pd
import geopandas as gpd
from osgeo import gdal
import glob
import numpy as np
import rasterio
import rasterio.mask
from os.path import exists
from shutil import copyfile
from pathlib import Path

#### Step 1: set up basic folders

def create_folder(name):
    try:
        os.mkdir(name)
    except FileExistsError:
        pass

create_folder('data') # folder for raw data
create_folder('shapefile')
create_folder('output') # folder for processed data
create_folder('output/drought')
create_folder('plots') # folder for plots
create_folder('stats') # folder for derived statistics
create_folder('stats/drought')
create_folder('stats/CCKP')
create_folder('stats/spei')
create_folder('maps') # folder for maps

country = os.getcwd().split('\\')[-1]

# Create corresponding folders on an external hard drive to store large data. Change file path as needed.
create_folder('F:/World Bank/City Scan/' + country)
create_folder('F:/World Bank/City Scan/' + country + '/data')
create_folder('F:/World Bank/City Scan/' + country + '/output')

#### Step 2: create AOI shapefiles
Have the AOI shapefiles in the shapefile folder before proceeding
The AOI shapefiles should be named such as new_york.shp (lower case, spaces replaced with underscores)

#### Step 3: project AOI shapefiles into WGS 84 and save them to city folders

cities = ['Kinshasa', 'Kananga', 'Mbuji-Mayi', 'Lubumbashi', 'Kisangani', 'Bukavu', 'Goma', 'Tshikapa', 'Mwene-Ditu', 'Gemena', 'Gbadolite', 'Matadi', 'Kikwit', 'Bunia']

for city in cities:
    create_folder(Path(city))
    create_folder(city / Path('maps'))
    create_folder(city / Path('data'))
    create_folder(city / Path('data') / 'AOI')
    create_folder(city / Path('stats'))

for city in cities:
    shp_4326 = gpd.read_file(Path('shapefile') / (city.replace(" ", "_").lower() + '.shp')).to_crs(epsg = 4326)
    shp_4326.to_file(Path(city) / 'data/AOI' / (city.replace(" ", "_").lower() + '_AOI.shp'))

#### Step 4: generate city coordinates and UTM

with open('centroids.csv', 'w') as f:
    f.write('city,x,y,utm\n')
    for city in cities:
        centroid = gpd.read_file(Path(city) / 'data/AOI' / (city.replace(" ", "_").lower() + '_AOI.shp')).centroid
        f.write('%s,%s,%s,%s\n'%(city, centroid.x[0], centroid.y[0], 32600+math.ceil((centroid.x[0]+180)/6)))

#### Step 5: extract country shapefile

# WB countries shapefile. Change file path as needed
all_countries = gpd.read_file(r"C:\Users\Owner\Documents\Career\World Bank\City Scan\data\wb_countries_admin0_10m\WB_countries_Admin0_10m.shp")

# change country name as needed
country_shp = all_countries[all_countries.NAME_EN == 'Democratic Republic of the Congo']

country_shp

country_shp.to_file(Path('shapefile') / (country.lower().replace(' ', '_') + '.shp'))

#### Step 6: make centroids shapefile

centroids = pd.read_csv('centroids.csv')

gdf = gpd.GeoDataFrame(centroids, 
                       geometry = gpd.points_from_xy(centroids.x, centroids.y))

gdf.to_file('shapefile/centroids.shp', crs = 'EPSG:4326')

#### Step 7: copy GEE outputs to city output folders

def crop_gee_output(city, raster_name):
    with rasterio.open(Path('output') / 'GEE' / (raster_name + '.tif')) as src:
        array = src.read(1)
        out_meta = src.meta.copy()
        out_meta.update({'nodata': 'nan'})
    with rasterio.open(city / Path('data') / (raster_name + '.tif'), 'w', **out_meta) as dest:
        dest.write(array, 1)

for city in cities:
    crop_gee_output(city, city.replace(" ", "_").lower() + '_Summer')

#### Step 8: copy files needed for Match Map Style toolbox

fathom_folder = Path('data') / 'Fathom'
floods = {'fluvial_undefended': 'FU',
          'pluvial': 'P'}
rps = [5, 10, 20, 50, 75, 100, 200, 250, 500, 1000]

for city in cities:
    for flood in floods:
        for rp in rps:
            copyfile(fathom_folder / flood / (floods[flood] + '_1in' + str(rp) + '.tif'), 
                     city / Path('data') / (floods[flood] + '_1in' + str(rp) + '.tif'))

#### Step 9: delete files needed for Match Map Style toolbox after maps are produced

for city in cities:
    for flood in floods:
        for rp in rps:
            os.remove(city / Path('data') / (floods[flood] + '_1in' + str(rp) + '.tif'))
