==> 20_cckp.ipynb <==
# processes climate projection data from CCKP

import xarray as xr
import pandas as pd
import numpy as np
from pathlib import Path
import glob
import os

data_folder = Path('F:/World Bank/City Scan/data/CCKP')  # change file path as needed

ssps = ['119', '245', '370']
periods = ['2040-2059']
varias0 = ['tas', 'txx', 'pr', 'r95ptot', 'cdd']
varias1 = ['hd35', 'tr26', 'wsdi', 'r20mm', 'r50mm']
varias = varias0 + varias1
rps = ['20yr', '50yr']

centroids = pd.read_csv('centroids.csv')

centroids

abs_val = dict({s: dict({p: dict({v: {} for v in varias + ['spei12']}) for p in periods}) for s in ssps})
ano_val = dict({s: dict({p: dict({v: {} for v in varias}) for p in periods}) for s in ssps})

# processing for variables without timedelta
for ssp in ssps:
    for period in periods:
        for varia in varias0:
            clim = xr.open_dataset(data_folder / ('climatology-'+varia+'-annual-mean_cmip6_annual_all-regridded-bct-ssp'+ssp+'-climatology_median_'+period+'.nc'))
            anom = xr.open_dataset(data_folder / ('anomaly-'+varia+'-annual-mean_cmip6_annual_all-regridded-bct-ssp'+ssp+'-climatology_median_'+period+'.nc'))
            
            for index, row in centroids.iterrows():
                x_coords = [row['x']-1, row['x'], row['x']+1]
                y_coords = [row['y']-1, row['y'], row['y']+1]
                clim_val_coords = []
                clim_val_vals = []
                anom_val_coords = []
                anom_val_vals = []
                
                for x in x_coords:
                    for y in y_coords:
                        clim_val_df = clim.sel(lon = x, lat = y, method = 'nearest').to_dataframe()
                        anom_val_df = anom.sel(lon = x, lat = y, method = 'nearest').to_dataframe()
                        if not (clim_val_df['lon'][0], clim_val_df['lat'][0]) in clim_val_coords:
                            clim_val_coords.append((clim_val_df['lon'][0], clim_val_df['lat'][0]))
                            clim_val_vals.append(clim_val_df['climatology-'+varia+'-annual-mean'].mean())
                        if not (anom_val_df['lon'][0], anom_val_df['lat'][0]) in anom_val_coords:
                            anom_val_coords.append((anom_val_df['lon'][0], anom_val_df['lat'][0]))
                            anom_val_vals.append(anom_val_df['anomaly-'+varia+'-annual-mean'].mean())
                
                clim_val = np.nanmean(clim_val_vals)
                anom_val = np.nanmean(anom_val_vals)
                
                abs_val[ssp][period][varia][row['city']] = clim_val
                ano_val[ssp][period][varia][row['city']] = anom_val

# processing for variables with timedelta
for ssp in ssps:
    for period in periods:
        for varia in varias1:
            clim = xr.open_dataset(data_folder / ('climatology-'+varia+'-annual-mean_cmip6_annual_all-regridded-bct-ssp'+ssp+'-climatology_median_'+period+'.nc'))
            anom = xr.open_dataset(data_folder / ('anomaly-'+varia+'-annual-mean_cmip6_annual_all-regridded-bct-ssp'+ssp+'-climatology_median_'+period+'.nc'))
            
            for index, row in centroids.iterrows():
                x_coords = [row['x']-1, row['x'], row['x']+1]
                y_coords = [row['y']-1, row['y'], row['y']+1]
                clim_val_coords = []
                clim_val_vals = []
                anom_val_coords = []
                anom_val_vals = []
                
                for x in x_coords:
                    for y in y_coords:
                        clim_val_df = clim.sel(lon = x, lat = y, method = 'nearest').to_dataframe()
                        anom_val_df = anom.sel(lon = x, lat = y, method = 'nearest').to_dataframe()
                        if not (clim_val_df['lon'][0], clim_val_df['lat'][0]) in clim_val_coords:
                            clim_val_coords.append((clim_val_df['lon'][0], clim_val_df['lat'][0]))
                            clim_val_vals.append(np.timedelta64(clim_val_df['climatology-'+varia+'-annual-mean'].mean()).astype('timedelta64[D]') / np.timedelta64(1, 'D'))
                        if not (anom_val_df['lon'][0], anom_val_df['lat'][0]) in anom_val_coords:
                            anom_val_coords.append((anom_val_df['lon'][0], anom_val_df['lat'][0]))
                            anom_val_vals.append(np.timedelta64(anom_val_df['anomaly-'+varia+'-annual-mean'].mean()).astype('timedelta64[D]') / np.timedelta64(1, 'D'))
                
                clim_val = np.nanmean(clim_val_vals)
                anom_val = np.nanmean(anom_val_vals)
                
                abs_val[ssp][period][varia][row['city']] = clim_val
                ano_val[ssp][period][varia][row['city']] = anom_val

# processing for spei12
varia = 'spei12'
for ssp in ssps:
    for period in periods:
        clim = xr.open_dataset(data_folder / ('climatology-'+varia+'-annual-mean_cmip6_annual_all-regridded-bct-ssp'+ssp+'-climatology_median_'+period+'.nc'))

        for index, row in centroids.iterrows():
                x_coords = [row['x']-1, row['x'], row['x']+1]
                y_coords = [row['y']-1, row['y'], row['y']+1]
                clim_val_coords = []
                clim_val_vals = []
                
                for x in x_coords:
                    for y in y_coords:
                        clim_val_df = clim.sel(lon = x, lat = y, method = 'nearest').to_dataframe()
                        if not (clim_val_df['lon'][0], clim_val_df['lat'][0]) in clim_val_coords:
                            clim_val_coords.append((clim_val_df['lon'][0], clim_val_df['lat'][0]))
                            clim_val_vals.append(clim_val_df['climatology-'+varia+'-annual-mean'].mean())
                
                clim_val = np.nanmean(clim_val_vals)
                abs_val[ssp][period][varia][row['city']] = clim_val

# write to csv
for ssp in ssps:
    for varia in varias + ['spei12']:
        with open('stats/CCKP/clim_'+varia+'_ssp'+ssp+'.csv', 'w') as f:
            f.write('city,'+periods[0]+'\n')
            for city in centroids.city:
                f.write("%s,%s\n"%(city, abs_val[ssp][periods[0]][varia][city]))
        if varia != 'spei12':
            with open('stats/CCKP/anom_'+varia+'_ssp'+ssp+'.csv', 'w') as f:
                f.write('city,'+periods[0]+'\n')
                for city in centroids.city:
                    f.write("%s,%s\n"%(city, ano_val[ssp][periods[0]][varia][city]))

aep_val = dict({s: dict({r: {} for r in rps}) for s in ssps})

# processing for future return period change factor
for ssp in ssps:
    for period in ['2035-2064']:
        for rp in rps:
            aep = xr.open_dataset(data_folder / ('changefactorfaep'+rp+'-rx5day-period-mean_cmip6_period_all-regridded-bct-ssp'+ssp+'-climatology_median_'+period+'.nc'))
            
            for index, row in centroids.iterrows():
                x_coords = [row['x']-1, row['x'], row['x']+1]
                y_coords = [row['y']-1, row['y'], row['y']+1]
                aep_val_coords = []
                aep_val_vals = []
                
                for x in x_coords:
                    for y in y_coords:
                        aep_val_df = aep.sel(lon = x, lat = y, method = 'nearest').to_dataframe()
                        if not (aep_val_df['lon'][0], aep_val_df['lat'][0]) in aep_val_coords:
                            aep_val_coords.append((aep_val_df['lon'][0], aep_val_df['lat'][0]))
                            aep_val_vals.append(aep_val_df['changefactorfaep'+rp+'-rx5day-period-mean'].mean())
                
                aep_val1 = np.nanmean(aep_val_vals)
                
                aep_val[ssp][rp][row['city']] = aep_val1

# write to csv
for ssp in ssps:
    for rp in rps:
        with open('stats/CCKP/aep_'+rp+'_ssp'+ssp+'.csv', 'w') as f:
            f.write('city,'+'2035-2064'+'\n')
            for city in centroids.city:
                f.write("%s,%s\n"%(city, aep_val[ssp][rp][city]))
