==> 21_spei.ipynb <==
# processes historical SPEI data

import xarray as xr
import pandas as pd
import numpy as np
from pathlib import Path
import glob
import os

data_folder = Path(r'F:\World Bank\City Scan\data\SPEI')  # change file path as needed

periods = ['01', '12', '48']
years = range(2011, 2021)

centroids = pd.read_csv('centroids.csv')

centroids

spei_val = dict({p: dict({c: {} for c in centroids.city}) for p in periods})

for period in periods:
    spei_nc = xr.open_dataset(data_folder / ('spei'+period+'.nc'))

    for index, row in centroids.iterrows():
        for year in years:
            for month in range(1, 13):
                time1 = str(year) + '-' + str(month) + '-15'
                val = spei_nc.sel(lon = row['x'], lat = row['y'], time = time1, method = 'nearest')['spei'].to_dict()['data']
                spei_val[period][row['city']][time1] = val

# write to csv
for period in periods:
    with open('stats/spei/spei'+period+'.csv', 'w') as f:
        f.write('city,date,spei\n')
        for city in centroids.city:
            for year in years:
                for month in range(1, 13):
                    time1 = str(year) + '-' + str(month) + '-15'
                    f.write('%s,%s,%s\n' % (city, time1, spei_val[period][city][time1]))
