==> 40_stats.ipynb <==
import pandas as pd
import geopandas as gpd
import rasterio
from rasterio.plot import show
import rasterio.mask
from rasterio import features
# import rasterstats
import numpy as np
import csv
import fiona
from pathlib import Path

city_list = pd.read_csv('centroids.csv').city

### AVERAGE TEMPERATURE

def avg_temp(city):
    city_nospace = city.replace(" ", "_").lower()
    temp_file = city / Path('data') / (city_nospace + '_Summer.tif')
    temp = rasterio.open(temp_file)
    temp_array = temp.read(1)
    return np.nanmean(temp_array)

cities_avg_temp = {}
for city in city_list:
    cities_avg_temp[city] = avg_temp(city)

cities_avg_temp

with open('stats/avg_temp.csv', 'w') as f:
    f.write('city,avg\n')
    for city in cities_avg_temp.keys():
        f.write("%s,%s\n"%(city, cities_avg_temp[city]))

### AIR QUALITY

def avg_air(city, year):
    city_nospace = city.replace(" ", "_").lower()
    temp_file = city / Path('data') / (city_nospace + '_air_quality_' + str(year) + '.tif')
    temp = rasterio.open(temp_file)
    temp_array = temp.read(1)
    temp_array = temp_array[temp_array >= 0]
    return np.nanmean(temp_array)

def bad_air(city, threshold = 5):  # threshold for the definition of "bad air"; default is 5 ug/m3
    city_nospace = city.replace(" ", "_").lower()
    air_file = city / Path('data') / (city_nospace + '_air_quality_' + str(year) + '.tif')
    air = rasterio.open(air_file)
    air_array = air.read(1)
    return sum(sum(air_array >= threshold)) / sum(sum(air_array != air.meta['nodata']))

cities_avg_air = {}
for city in city_list:
    cities_avg_air[city] = {}
    for year in range(1998, 2020):
        cities_avg_air[city][year] = avg_air(city, year)

cities_bad_air = {}
for city in city_list:
    cities_bad_air[city] = {}
    for year in range(1998, 2020):
        cities_bad_air[city][year] = bad_air(city)

# cities_avg_air

with open('stats/avg_air_1998_2019.csv', 'w') as f:
    f.write('city,year,avg,pct_bad_air\n')
    for city in cities_avg_air.keys():
        for year in range(1998, 2020):
            f.write("%s,%s,%s,%s\n"%(city, year, cities_avg_air[city][year], cities_bad_air[city][year]))

### LANDSLIDE AVERAGE FREQUENCY

def landslide(city):
    city_nospace = city.replace(" ", "_").lower()
    ls_file = city / Path('data') / (city_nospace + '_landslide.tif')
    ls = rasterio.open(ls_file)
    ls_array = ls.read(1)
    return np.nanmean(ls_array)

cities_ls = {}
for city in city_list:
    try:
        landslide_value = landslide(city)
        cities_ls[city] = landslide_value
    except:
        pass

cities_ls

with open('stats/landslide_avg.csv', 'w') as f:
    f.write('city,avg\n')
    for city in cities_ls.keys():
        f.write("%s,%s\n"%(city, cities_ls[city]))
